+++ release-notes.md
@@ -13,6 +13,7 @@ file's top entry doesn't match the pushed tag. See ADR 0002.
 - Add preview for OBJ and MTL formats
 - Add preview for 3MF format
 - Add preview for STEP format
+- Load files on background thread
 
 ## v0.2.0 - 2026-08-08